The work is almost done...
The Computer Graphics subject is almost finished, but we have only implemented half of the features. We need to improve our work, but that's a story from another time. However, it's important to note that the revolution isn't finished yet. On the contrary, the revolution is just beginning. There are still many exciting advancements and innovations ahead in the world of Computer Graphics.